Re-write in Python
This is a much more straightforward language in which to write this
utility.

The Python script is functional, and can rename files in the majority
of cases. However, there are many cases in which the script will fail
e.g. if one of the files has a name corresponding to an enumerated
file. These issues will be dealt with in later commits.